+++ NEWS.md
@@ -1,5 +1,8 @@
-# teal.osprey development version
-- to edit
+# teal.osprey 0.1.2.9000
+
+# teal.osprey 0.1.2
+- New waterfall module
+- Refactored due to recent changes in `utils.nest`
 
 # teal.osprey 0.1.1
 Refactoring package according to the NEST standards.